+++ lib/src/components/components.dart
@@ -4,13 +4,17 @@
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-// http://www.apache.org/licenses/LICENSE-2.
+// http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
+//
+// This file contains code adapted and ported from the open-source
+// Flowbite library (Copyright (c) Bergside Srl), which is licensed
+// under the MIT License.
 
 export 'flowbite_accordion.dart';
 export 'flowbite_alert.dart';
@@ -18,6 +22,7 @@ export 'flowbite_avatar.dart';
 export 'flowbite_avatar_group.dart';
 export 'flowbite_avatar_group_label.dart';
 export 'flowbite_badge.dart';
+export 'flowbite_bottom_navigation.dart';
 export 'flowbite_button.dart';
 export 'flowbite_button_group.dart';
 export 'flowbite_icon_shape.dart';
@@ -25,3 +30,4 @@ export 'flowbite_input_field.dart';
 export 'flowbite_input_label.dart';
 export 'flowbite_outlined_button.dart';
 export 'flowbite_spinner.dart';
+export 'flowbite_toggle_input.dart';