+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Add `FlowbiteTheme.extension` static method.
+- Add `FlowbiteBottomNavigation` component.
+- Add `FlowbiteToggleInput` component (Advanced type).
 
 ### Changed
 - Migrate to `material_ui`.